refactor(core): split band guide types, label layout, and planner
Extract band-guide-types and band-label-layout (wrap/cap helpers) from
the measured planBandAxis planner; keep band-guide.ts as the stable
facade. Split characterization into tests/band-guide/ suites by
escalation, rotation/margin, cardinality, and contracts.
