+++ packages/core/src/layout/band-guide-types.ts
@@ -1,0 +1,67 @@
+/**
+ * Band axis guide plan types — shared by the planner and guide-plan contracts.
+ */
+import type { TextMeasurer } from "./measure.js";
+
+export type BandLabelMode = "single-line" | "wrapped" | "rotated";
+
+/** A pre-resolved band tick to lay out (break-matching already applied). */
+interface BandPlanEntry {
+  value: string | number;
+  label: string;
+  domainIndex: number;
+}
+
+export interface BandAxisPlanInput {
+  aesthetic: "x" | "y";
+  panelIndex: number;
+  /** Total number of bands (band width = extentPx / categoryCount). */
+  categoryCount: number;
+  /** Ticks to lay out, in display order, already break-filtered. */
+  entries: readonly BandPlanEntry[];
+  /** Only "horizontal" is planned; "vertical" returns a single-line no-op plan. */
+  orient: "horizontal" | "vertical";
+  extentPx: number;
+  reverse: boolean;
+  measurer: TextMeasurer;
+  fontSize: number;
+  /** Cap for along-axis label overhang past the end ticks (right/left margin). */
+  marginCapPx: number;
+  /** Cap for the orthogonal label band (bottom margin for x). */
+  orthogonalMarginCapPx: number;
+  /** Margin quantum; fitness budgets are floored to it so mode is A→B stable. */
+  quantum?: number;
+  ellipsis?: string;
+  /** Pass-A mode; the planner escalates only, never de-escalates. */
+  previousMode?: BandLabelMode | null;
+}
+
+export interface BandAxisPlanTick {
+  value: string | number;
+  label: string;
+  fullLabel: string;
+  labeled: boolean;
+  domainIndex: number;
+  /** Wrapped lines (present when mode === "wrapped"). */
+  lines?: string[];
+  /** Rotation in degrees (present when mode === "rotated"). */
+  angle?: number;
+}
+
+export interface BandAxisPlan {
+  mode: BandLabelMode;
+  /** 0 | -45 | -90 */
+  angle: number;
+  ticks: BandAxisPlanTick[];
+  /** 1 = every category labeled; >1 = high-cardinality thinning. */
+  labelEvery: number;
+  /** Orthogonal (bottom) band height the labels require, px. */
+  labelBandHeight: number;
+  /** Along-axis overhang past the RIGHT end tick (into the right margin), px. */
+  alongOverhang: number;
+  /** Along-axis overhang past the LEFT end tick (into the left margin), px. */
+  leftOverhang: number;
+  overlap: boolean;
+  marginOverflow: boolean;
+  degraded: string[];
+}
